AI-EMAIL-ASSISTANT-main/backend_fastapi/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
motor>=3.1.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
aiofiles>=23.1.0
celery>=5.3.0
redis>=4.5.0
transformers>=4.30.0
torch>=2.0.0
openai>=0.27.0
email-validator>=1.3.1
pydantic[email]>=1.10.10
scikit-learn>=1.2.2
joblib>=1.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0
boto3>=1.26.0
pytest>=7.0.0